+++ polls/test/test_question_index.py
@@ -1,0 +1,91 @@
+"""Add unittest for KU-polls."""
+import datetime
+from django.test import TestCase
+from django.utils import timezone
+from polls.models import Question
+from django.urls import reverse
+
+
+def create_question(question_text, days, end_date=30):
+    """
+    * copy comment from tutorial.
+
+    Create a question with the given `question_text` and published the
+    given number of `days` offset to now (negative for questions published
+    in the past, positive for questions that have yet to be published).
+    """
+    publish_time = timezone.now() + datetime.timedelta(days=days)
+    # add ending date
+    ending_time = publish_time + datetime.timedelta(days=end_date)
+    return Question.objects.create(question_text=question_text,
+                                   pub_date=publish_time, end_date=ending_time)
+
+
+class QuestionIndexViewTests(TestCase):
+    """unittest for redirect page."""
+
+    def test_no_questions(self):
+        """
+        * copy comment from tutorial.
+
+        If no questions exist, an appropriate message is displayed.
+        """
+        response = self.client.get(reverse('polls:index'))
+        self.assertEqual(response.status_code, 200)
+        self.assertContains(response, "No polls are available.")
+        self.assertQuerysetEqual(response.context['latest_question_list'], [])
+
+    def test_past_question(self):
+        """
+        * copy comment from tutorial.
+
+        Questions with a pub_date in the past are displayed on the
+        index page.
+        """
+        create_question(question_text="Past question.", days=- 30)
+        response = self.client.get(reverse('polls:index'))
+        self.assertQuerysetEqual(
+            response.context['latest_question_list'],
+            ['<Question: Past question.>']
+        )
+
+    def test_future_question(self):
+        """
+        * copy comment from tutorial.
+
+        Questions with a pub_date in the future aren't displayed on
+        the index page.
+        """
+        create_question(question_text="Future question.", days=30)
+        response = self.client.get(reverse('polls:index'))
+        self.assertContains(response, "No polls are available.")
+        self.assertQuerysetEqual(response.context['latest_question_list'], [])
+
+    def test_future_question_and_past_question(self):
+        """
+        * copy comment from tutorial.
+
+        Even if both past and future questions exist, only past questions
+        are displayed.
+        """
+        create_question(question_text="Past question.", days=- 30)
+        create_question(question_text="Future question.", days=30)
+        response = self.client.get(reverse('polls:index'))
+        self.assertQuerysetEqual(
+            response.context['latest_question_list'],
+            ['<Question: Past question.>']
+        )
+
+    def test_two_past_questions(self):
+        """
+        * copy comment from tutorial.
+
+        The questions index page may display multiple questions.
+        """
+        create_question(question_text="Past question 1.", days=- 30)
+        create_question(question_text="Past question 2.", days=- 5)
+        response = self.client.get(reverse('polls:index'))
+        self.assertQuerysetEqual(
+            response.context['latest_question_list'],
+            ['<Question: Past question 2.>', '<Question: Past question 1.>']
+        )
